Revert Commit: Move Test under tests/ folder (#429)
